+++ game/healthItem.go
@@ -1,28 +1,31 @@
 package game
 
 import (
-	"liner/dataio"
+	"image/color"
+	
 
-	"github.com/gen2brain/raylib-go/raylib" 
+	"github.com/gen2brain/raylib-go/raylib"
 )
 
 type HealthTile struct {
-	BorderTile rl.Rectangle
+	//BorderTile rl.Rectangle
 	HealthItem rl.Rectangle
 	IsHidden bool
+	Color color.RGBA
 }
 
 func (t *HealthTile) Draw() {
-	rl.DrawRectangleRec(t.BorderTile, dataio.Palette["lightgray"])
+	//rl.DrawRectangleRec(t.BorderTile, dataio.Palette["lightgray"])
 	if !t.IsHidden {
-		rl.DrawRectangleRec(t.HealthItem, dataio.Palette["red"])
+		rl.DrawRectangleRec(t.HealthItem, t.Color)
 	}
 }
 
 func NewHealthTile(posx int32, posy int32) *HealthTile {
 	return &HealthTile{
-		BorderTile: rl.NewRectangle(float32(posx), float32(posy), 25,25),
-		HealthItem: rl.NewRectangle(float32(posx)+5, float32(posy)+5, 15,15),
+		//BorderTile: rl.NewRectangle(float32(posx), float32(posy), 25,25),
+		HealthItem: rl.NewRectangle(float32(posx), float32(posy), 25,25),
 		IsHidden: true,
+		Color: rl.Red,
 	}
 }